+++ BitChopp.sln
@@ -3,7 +3,7 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
 VisualStudioVersion = 17.5.002.0
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "BitChopp", "BitChopp.csproj", "{0DC4ACCA-CBEC-4D57-A61C-5A249C8AB85F}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "BitChopp.Main", "src/BitChopp.Main/BitChopp.Main.csproj", "{0DC4ACCA-CBEC-4D57-A61C-5A249C8AB85F}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
