+++ src/BitChopp.Main/Assets/logo.png
