+++ src/BitChopp.Main/Views/MainWindow.axaml
@@ -2,10 +2,10 @@
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
         xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
         xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-        xmlns:vm="clr-namespace:BitChopp.ViewModels;assembly=BitChopp"
+        xmlns:vm="clr-namespace:BitChopp.Main.ViewModels;assembly=BitChopp.Main"
         x:DataType="vm:MainViewModel"
         mc:Ignorable="d" d:DesignWidth="600" d:DesignHeight="1024"
-        x:Class="BitChopp.Views.MainWindow"
+        x:Class="BitChopp.Main.Views.MainWindow"
         Title="BitChopp - Casa21"
         Background="Black"
         WindowStartupLocation="CenterScreen">
@@ -32,7 +32,7 @@
 
   <StackPanel HorizontalAlignment="Center" VerticalAlignment="Center">
     <!-- Logo -->
-    <Image Source="avares://BitChopp/Assets/logo.png" Width="200" Height="200"/>
+    <Image Source="avares://BitChopp.Main/Assets/logo.png" Width="200" Height="200"/>
     
     <!-- Glass Size Buttons -->
     <ItemsControl ItemsSource="{Binding Switches}">
