+++ src/BitChopp.Main/Views/QRCodeWindow.axaml
@@ -3,7 +3,7 @@
         xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
         xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
         mc:Ignorable="d" d:DesignWidth="600" d:DesignHeight="1024"
-        x:Class="BitChopp.Views.QRCodeWindow"
+        x:Class="BitChopp.Main.Views.QRCodeWindow"
         Title="BitChopp - Casa21 - Pay!"
         Background="Black"
         WindowStartupLocation="CenterScreen">
